pyproject.toml
[tool.poetry]
name = "message-board"
version = "0.1.0"
description = ""
authors = ["Nyaaa <1457047+Nyaaa@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
django-filter = "^23.1"
python-dotenv = "^1.0.0"
django-modeltranslation = "^0.18.9"
django-widget-tweaks = "^1.4.12"


[tool.poetry.group.dev.dependencies]
django-stubs = "^1.16.0"
coverage = "^7.2.2"
model-bakery = "^1.10.1"
faker = "^18.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
